Fixes for linux build
* fixed capitalization in CMakeList.txt
* use unsigned long instead of size_t in image.i

+++ include/image.h
@@ -26,7 +26,7 @@ protected:
 
 public:
 
-    std::vector<size_t > get_tttr_indices(){
+    std::vector<unsigned long > get_tttr_indices(){
         return tttr_indices;
     }
